pyproject.toml
[project]
name = "taskbeacon-mcp"
version = "0.1.8"
description = "A model contexture protocal (MCP) for TaskBeacon"
authors = [
  { name = "Zhipeng Cao", email = "zhipeng30@foxmail.com" }
]
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
license-files = ["LICENSE"]
dependencies = [
  "httpx",
  "GitPython",
  "mcp[cli]",
  "ruamel.yaml",
  "edge_tts",
  # Bugfix: the code imports `fuzzywuzzy`, not `fuzzywuzz`.
  "fuzzywuzzy",
]
keywords = ["taskbeacon", "mcp", "uv", "TaskBeacon"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Operating System :: OS Independent",
]

[project.urls]
Homepage = "https://github.com/TaskBeacon/taskbeacon-mcp"
Repository = "https://github.com/TaskBeacon/taskbeacon-mcp"
"Bug Tracker" = "https://github.com/TaskBeacon/taskbeacon-mcp/issues"

[project.scripts]
taskbeacon-mcp = "taskbeacon_mcp.server:main"

[project.optional-dependencies]
dev = [
  "pytest",
]


[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.uv]
# This belongs here, not under [project]
required-version = ">=0.8.3"

[tool.setuptools]
packages = { find = { exclude = ["tests*"] } }

[tool.setuptools.package-data]
taskbeacon_mcp = ["prompts/*.md"]
